:sparkles: Feat: Seperate concerns for pilot landing page
